# Message Pack optimizations