Some work mostly around faster delegate invocation with fewer allocations.

Before:
|                           Type |  Method |        Job | UnrollFactor |         Mean |     Error |    StdDev |       Median |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------------------------------- |-------- |----------- |------------- |-------------:|----------:|----------:|-------------:|-----------:|-----------:|----------:|----------:|
|       IntMaybeMapSerialization | MsgPack | DefaultJob |           16 |    15.16 ms |  0.219 ms | 0.205 ms |    15.08 ms |   984.3750 |   281.2500 |   31.2500 |   8.55 MB |
|   RecursiveRecordSerialization | MsgPack | DefaultJob |           16 |   518.65 ms |  1.895 ms | 1.680 ms |   519.25 ms | 19000.0000 |          - |         - | 173.21 MB |
| RecursiveRecordDeserialization | MsgPack | Job-VVPEKY |            1 |   541.63 ms |  3.090 ms | 2.413 ms |   540.80 ms | 26000.0000 |  9000.0000 |         - | 209.36 MB |

Now:
|                           Type |  Method |        Job | UnrollFactor |         Mean |     Error |    StdDev |       Median |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------------------------------- |-------- |----------- |------------- |-------------:|----------:|----------:|-------------:|-----------:|-----------:|----------:|----------:|
|       IntMaybeMapSerialization | MsgPack | DefaultJob |           16 |    11.973 ms | 0.0418 ms | 0.0371 ms |    11.959 ms |   859.3750 |    46.8750 |         - |   7.29 MB |
|   RecursiveRecordSerialization | MsgPack | DefaultJob |           16 |   399.717 ms | 1.3956 ms | 1.3055 ms |   399.781 ms | 14000.0000 |          - |         - |  128.5 MB |
| RecursiveRecordDeserialization | MsgPack | Job-JLAASO |            1 |   427.218 ms | 2.9600 ms | 2.7688 ms |   426.836 ms | 20000.0000 |  5000.0000 |         - | 160.93 MB |