@Zaid-Ajaj 2020-07-05
More optimizations, more better 😍 merged and published :rocket: the relevant packages (everything except for Fable.Remoting.Json) and they should be available soon after nuget does its magic